rename : 파일 구조 수정 및 이동
