+++ second-seminar/src/main/java/com/server/dosopt/secondseminar/dto/request/MemberCreateRequest.java
@@ -1,6 +1,6 @@
 package com.server.dosopt.secondseminar.dto.request;
 
-import com.server.dosopt.secondseminar.SOPT;
+import com.server.dosopt.secondseminar.domain.SOPT;
 import lombok.Data;
 
 @Data
