+++ second-seminar/src/main/java/com/server/dosopt/secondseminar/dto/request/MemberProfileUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.server.dosopt.secondseminar.dto.request;
 
-import com.server.dosopt.secondseminar.Part;
+import com.server.dosopt.secondseminar.enums.Part;
 import lombok.Data;
 
 @Data
